Adds robots.txt file to prevent search engine indexing
This repo is intended to be a demo site based off of The Lovely Dev's
own website. It will showcase the benefits of building websites with
Astro. In order to avoid confusion, we don't want this site to be
indexed by search engines. The robot.txt file will do this for us.